+++ CHANGELOG.md
@@ -1,11 +1,9 @@
 ## 0.0.3 (unreleased)
 
-
-- Nothing changed yet.
+- persist and restore state on i3 restarts
 
 
 ## 0.0.2 (2025-02-05)
-
 
 - fix output_blacklist config logic
 - add screenshot_lib_path config item